use accordions for all content box sections (#45)
Adds a simple accordion component and uses that for all content box sections and the related-buildings.
I moved around styles for better separation between pages and components.
Moved all the know-your-rights content boxes to a separate file to simplify the results file and since we'll reuse in standalone page. Also changed them all to components to easily change the header text based on context.