refactor results styles, update print styles, expand accordions (#55)
Updates the styles for printing the results page (guide pages to follow in separate PR)

This ones has a lot of reorganization of styles (lots of moving around, and changed nested), and also some minor additions to component (eg. adding print dividers after content boxes) that created big diffs where there aren't real changes.

All closed accordions are automatically opened for the print and then reset back to preexisting state afterwards.

I had to make some changes from the design files to keep everything on the page. In particular, the text was too big on the print page, so I also changed the base font size when printing down to 12px and that helped. Will review with Corey, but good to merge this in first since there are so many changes it'll make it hard to do other things without lots of merge conflicts.

Here's an example of a print ed result: gce-result-print.pdf

might add page-break-after: always; on contentbox, waiting for corey feedback