sort related properties (#70)
For the sorting, I've added some addition criteria to sort by:

whether there are acris doc links,
number of units in the property compared to how many additional units the user needs to find,
whether the property shares the exact same ownername in DOF records
whether the property is listed alongside the search address in a multi-BBL acris document
Also, in working on this I was reminded of soe annoying type issues with the WOW api data so made a small adjustment (mostly on the wow side JustFixNYC/who-owns-what@08acf27). It was confusing because there were NULLs instead of arrays when there was no data, but now it'll be empty arrays.