address search bar highlights existing text when focused (#77)
When address bar has text in it already and is focused it highlights the existing text to make it clear that if you start typing it will clear all of that. react-select makes it really tricky to actually access the input element to be able to use input.select() to actually select it, but this visual highlighting works.