add privacy and terms of use pages (#85)
This copies over our privacy policy and terms of use pages from WOW. I follow the same set up as WOW - where we download static copies of the contentful pages and use contentful packages to render the elements. I made only a couple small edits to the data files to adjust the title and links. I didn't do a lot with the styles, but just adjusted some of the heading styles since these pages use a mix of many header levels.
One issue is that the contentful data uses headings up to level 1, but our own page titles are 2, so this breaks some a11y rules. I think the only solution would be to better standardize things on contentful to start at 3, but not sure this is necessary.