add github action to scan for deadlinks (#101)
adds our github action that checks for dead links.

I continued to have problems with good links being marked as broken, and after looking more into it is seemed like other people were having problems as well and the project haven't been updated in many years now. So I switched over to a different package for the link scanning and updated the repo and tested it out on the orgsite, and it seems to be working better.

https://github.com/JustFixNYC/deadlink-crawler
JustFixNYC/justfix-website#396

I adjusted the formatting of the GH issue to show it two ways - one like it was with each page and the broken links on that page, and an alternative format with each unique broken url and all the pages it appears on (this is helpful when it's something like the generic rent stab fact sheets that is linked across many LC articles)