Add Subsidy guide page and related content (#102)
This adds a new result type for subsidized, and a guide page to determine subsidy status, and updates the criteria logic and survey question helper text for subsidies.

At the same time I also refactored the JFCL link to simplify things a bit. Nows it's all one file and there is a generic version with internal and external versions. This was needed to add a different icon for the link to the subsidy guide from the question helper text. I also added the new icon to the component library: JustFixNYC/component-library#54. (this has now been cut from designs, but I think the code clean up is still helpful)

We're still waiting on a final copy, but that can be done later as part of the site-wide copy update.

This also include a small update to the result page criteria for RS when our data contradicts the user selection (they say unsure/no but all units are registered or it's 421a/j51), in the same style as we now do for the subsidy criteria