show/hide header on scroll (#106)
When scrolling down the topbar should hide, and show on scroll up.

Adds two hooks - to detect scroll direction, and detect if element is in the viewport. The header gets position sticky, and a placeholder div remains at the top to keep the space. when scrolling down the sticky header dissapears, and reappears when scrolling up. When you get to the top and the invisible placeholder is visible, the sticky header always shows.

This setup seems to work ok, but the downside is that the hiding of the sticky header is done with display:none and so can't be transitioned to slide away. I briefly tried to just shift the sticky header off screen using top: -100px, but then it would glitch out at the top of the page