wait to send results to backend until ready (#120)
We got some errors from the tenants2 backend where the result criteria values were missing. It looks like this happens if you go directly to a result page url and it needs to fetch the building data first but the useEffect was sending before it was ready.

this PR makes two fixes for this. One is that the check for values we did before was incomplete - since the object will be there but the internal fields are all undefined. Then I just made this check a variable and added it to the dependencies for the useEffect so that if at first the data isn't ready and it skips the POST it will retry once the dataReady variable becomes true