remove prod condition for TP connection, send all data with results (#123)
We were having a problem when someone uses a results page link without a user id in the search params, we are only sending the results info to the tenant platform backend but we'll need to create a new record and are missing required fields (address fields). So we need to send all the data along with the results, and then also update the user in session storagewith the ID if there wasn't one set already.

I also removed a condition we had left over from early on that only sent data to the backend on production, but there's no need for that, since we also just catch any errors so when you're working locally on front-end only things you still don't need to be runnings a local tenants2 server.

There is also a companion PR on tenants2: JustFixNYC/tenants2#2477