encode url params, add results email copy (#125)
Adds the copy for the "email coverage" feature from result page. Because of the way mailto links work as urls, we had a problem of the urlencoded search params of the link included in the body copy getting decoded in the email, so the link breaks. So I borrowed some code from signature dashboard to compress/encode search params data to avoid this.

Waiting on confirmation about copy for the guide page emails

[sc-16294]