wrap inputs for enter-key submit on address, calculator, phone (#130)
We noticed that on all out inputs you can press enter to submit, but that was because we weren't actually using a form element for these - specifically the rent calculator, the address search on homepage, and the phone number submission component. I've updated all of these to just change existing container divs into forms and added preventDefault() so the page doesn't reload.