Update CPI and rent levels with new DHCR published numbers (#138)
The new CPI number has been published, and this updated to the new value: 3.79%

Sources:

DHCR published this fact sheet with the newest CPI: https://hcr.ny.gov/system/files/documents/2025/04/gce-fact-sheet-04.23.25-update.pdf#page=7
HPD's know your rights page uses the same new CPI, and says "as of February 19, 2025", so I included updated the date we had on the site as well.
This also makes use of variables for the copy on the rent calculator page, moves the CPI var to that page (feels like the logical place to set it now), and also adds a var for the date of the new CPI - this listed on HPD's website as February 19, 2025.

It turns out that in the law it was never specified when the new CPI/Rent numbers take effect, only that they must be published on/before Aug 1 each year. So for now we'll just keep this language of "as of February 19, 2025" that HPD is using, until later things are clarified.

I also updated the rent levels, which are also updated in the above DHCR document.