Add Lingui for internationalization, Spanish translations (#160)
This adds Lingui to support multiple languages. We are using the latest v5 version of lingui (ahead of old versions on wow and elsewhere), and the lingui vite plugin.

We are using Crowdin with github integration again to manage translations. We are also experimenting with using DeepL for machine translations on Crowdin for "pre-translations", then human translators are still translating everything but the pre-translations serve as template/first draft. For now we only have the machine translation, but will update as soon as the human translators are done.

So far we only have Spanish, but may add more languages in the future.

We use the same setup as in other projects with locale prefix on all routes and auto detection of system preferred language when landing on a url without prefix.

For now we have a simple en/es toggle UI in the footer to change languages. Later we will add a dropdown list from the nav menu to support multiple languages.