refactor rentstab criteria table copy to simplify lingui (#181)
We were having a strange problem with lingui where in the english version the entire section f copy was replaced with a few jumbled characters. The copy for the user value for rent stab in the criteria table was composed with a lot of conditional logic and embedded subsections (example), so this was probably breaking lingui's ability to piece together the strings properly. This PR refactors to simplify the logic for lingui and minimize embedded sections.

For some reason the original problem was not happening in local dev, only on the demo/prod site, so on this branch I also manually edited the es/messages.po to add in new translations and it's showing up as intended on the preview deploy link (I then reverted those edits).