Add letter reason steps (#185)
Adds step for selecting reason for the letter (planned rent increase or non renewal of lease), and a followup step to clarify the reason (if increase is above reasonable standard or if the landlord gave a good cause reason for non renewal).

I also added to the "back" button to clear values from the current step before going back, this prevents an awkward situation where if you land on a page (eg user details) then go back to change answers, and when you return to user details it shows error messages right away.

Unrelated, but I moved around the rent increase limit variables so exports on rent calculator script doesn't prevent fast-reloading.

Eventually the steps will need to be updated with the new system for managing the flow through steps and progress bar etc.

I updated the zod schema to use their discriminatedUnion method to allow for conditionally required fields based on the selected reason for the letter. This was a little awkward when working with the react-hook-form errors because it wasn't interpreting which branch you're on like is was supposed to, so I added some typescript typeguard predicate functions to fix this.

There are some changes we'll need to make for the non-renewal branch, since we are talking about maybe just an intersitital, so this separate step won't be needed. I figure we can just deal with all that in a separate PR to add that once designs are ready - but let me know if you prefer me to just remove all of that now.

In the process of doing this I also realized we need to fix our zod+lingui integration so that the errors messages are translated. But this PR was already getting big so I'll follow up with a separate one for that work.