add lingui to zod error messages (#190)
This PR adds lingui for internationalization of zod validation error messages. Because the zod schema is defined outside of react it doesn't get the context of the active i18n from context provider. It's also preferable to keep the schema defined separately, rather than within the letter react component, since the type inferred from the schema is referenced across all the separate form step components. I explored some solutions using the zod error mapping method, which would involve specifying all the errors with identifiers and create a separate map of error type to message, but i didn't like the idea of having to keep track of the validation rules and messages separately. Turns out the easiest solution was just to wrap all the zod schema definitions in a function, so that we can pass the i18n variable at the time it's used (in the react-hook-form with zodResolver), and then we only needed to adjust the type typescript type inference with ReturnType<>, and nothing else has to change!