add letter copy/layout, unify styles, add test sender (#193)
This adds all the components of the letter and copy (this will be updated later). I've also realized we don't really need the separate styles for html and pdf, so I've unified all of that. I've updated the letter preview step (designs will be updated soon for language toggle, etc.). I've also added a temporary component on the letter main page to preview and send a letter using our sample data to help with checking letter content/style changes and testing the backend by sending letter (eg to trigger textit flow)