update mail choice and emails step (#194)
Update the mail choice step to include user, landlord, and extra emails.

No styling has been applied to the step, so it looks rough, but we'll do that all at once in a separate PR.

I've temporarily reordered the steps so that this new mail choice one appears before landlord details which required tenants2 connection, so it's easier to test/review. Later we can sort all that out, probably along with the step management/progress bar work.

Also fixes a bug in the user address set p - use of react-hook-form's reset to set multiple values of user_details at once was creating weird issues for any subsequent step that also use fields from user_details where validation errors wouldn't be caught.

This also adds the field extra_emails, which need to be an array of objects (rather than a flat array) to work with react-hook-form's useFieldArray. And so before sending data to tenants2 it needs to be flattened.

I've also removed the required boolean field email_to_landlord, since we can just infer from the presence of the optional landlord email field. Companion PR on tenants2 removing it from the backend: JustFixNYC/tenants2#2487