add letter landing page (#197)
Adds the landing page for the letter builder.

For now I've updated the routes to use /letter for the landing page and /letter/build for the steps, but I think we'll be updating to have separate routes for each step latter with the progress bar/step management PR.

This also updates JFCL for the change to Font Awesome v7 to get some new icons including the mailbox/envelope one used for the main CTA here. As part of the JFCL change we removed the type prop for <Icon> instead adding the type as a suffix to icon names in the existing icon prop. For details see JustFixNYC/component-library#58

Now that #198 has been merged in, adding routes fro the letter steps, I've made some changes to fit in landing page with the new routes. So now /letter is the landing page, and the CTA to begin the survey navigates to /letter/reason. And I removed the header from since it's not in the current mocks and it also didn't work with the nested route/ system when adding the landing page. This is mostly just to get this PR merged in now, and we can sort out any changes to the routing later maybe when we prevent the direct linking.

Like with everything else I've left out the styles from this page as well.