Combine user contact and address steps, add no unit checkbox (#196)
This implements the final design for the user details step - including all user_details fields and removing the separate address step.

This also adds the "no unit number" checkbox and conditional zod validation logic with secondary_line, which requires an input for one or the other (exclusive). I tried to do this in a clean dry way, but it doesn't seem possible so the validation logic is repeated. So I also didn't go through with the nested address properties as I thought I would.

Also adds a "learn more" modal for the user email/phone section explaining why we ask for that from users.

To allow for the geosearch address autocomplete bar to be pre-filled with user inputs when navigating back to a completed step we needed to make it a controlled component.

As with other recent PRs this doesn't include any styling, but all the components are there and logic is working.