add context provider, update reason steps for modal (#199)
This PR changes the non-renewal reason step to override the "next" button to launch an interstitial modal giving a warning if the user selects that the landlord gave a "good cause" reason. This was previously a dead-end step like when the increase is "reasonable", so that step is deleted now.

To allow for this behavior of the "next" button not submitting the form we need to override the button. To make this easier to do and keep that logic with the step component, we now use a context provider to help manage state between the form builder and the steps.

The context makes the back and next functions from the builder available to children. This allows us to have a standalone where the buttons have the original behavior as their defaults (type="submit", etc) but can easily be overridden to hide a button or change any of it's props.

Now that we have the context set up, I also added all of the react-hook-form methods so that they don't need to be passed as props to every form step component.

All the existing steps have also been updated to use this method (excluding the WIP landlord details)

There are a few other small unrelated changes here as well:

add "(optional)" to some input labels, and separate it out to reduce repetition for translations
fix the optional validation for landlord email (looseOptional(), which pre-process to turn "" into undefined to distinguish between empty and invalid format)
move the letter submission tester over to a separate route so the full builder experience starts to match mocks