add cc_user form field and checkbox to mail_choice step (#204)
Adds the new cc_user boolean field to form validation types, and adds the checkbox to the mail_choice step.

The checkbox is disabled until a value is entered for landlord email.
If it is checked but no user email is entered an error is raised on the user email saying to either enter your email or uncheck the cc box.

There are a couple of temp fixes that we'll need to revisit:

if the box is checked but then you remove your landlord email the box is automatically unchecked and disabled. If you then reenter a landlord email the box is enabled and set back to checked automatically.
if you check the cc box then go back a step and remove your email from the contact info page then try to go "next" it would raise the error to enter and email or uncheck the box but you can't get to the box. so for now when you go "back" from the mail choice step the cc box is automatically reset to the default unchecked state. This goes against our general rule of not clearing user entered data on "back", so we should decide how to handle this. It seems like it's gonna be hard to condition the validation based on the steps as well.
Companion PR for backend: JustFixNYC/tenants2#2490