New step management system (#206)
This does a big refactor to use a different approach to managing steps in the letter builder. Originally we started with an array of steps and used state to keep track of an index to access the current step info. The major limitation of this is that it can't accommodate branching paths with different numbers of steps. Once we added the routes it became possible to know the current step just from the url pathname, and moving between steps became just a navigate() call to a new route. So this new approach makes use of that and puts the component into the step list as well. previously we needed to conditionally render steps based on the step index, and that gets messy. now we cam make further use of the context provider and move the logic for where to go for the next step into each step component, since whenever there is going to be branching it's usually based on responses in the previous step so it keeps that all together.

Now that the progress through steps isn't linear, I've just hardcoded progress bar values into the steps object as well, and that might actually be easier/more clear than overriding with conditional logic.

I also refactored the main builder component to remove the

and now that lives in each step, since not all steps in the builder are actually forms (eg. preview, confirmation, allowed-increase). This also helps with the next step progression handling, and again makes use of the context.