disable/clear unit input when no-unit checked (#207)
When the "no-unit" checkbox is checked the text input for unit is disabled, and if there is already a value it is cleared. This applies to user and landlord steps. I also clean up the defaultValues supplied to react-hook-form and include user primary address since the undefined value causes an unintended error message