remove setValue onBlur by not rendering form fields in modal until open (#209)
Because we need to show the same form fields either on the page or in the modal we had two copies. But I think the contents of the modal are still present in the dom when it's not open, so we had multiple inputs registered for the same fields and they would clash (sometimes overwriting values with ""). So now we only render the form fields in the modal once it's open and that seems to solves the issue. I believe this was the reason we were using the setValue onBlur for all the inputs, so I've removed that and everything is working still.