use onsubmit instead of type="button" override, add step classname to form (#215)
We sometimes need to override the default behavior of "next" button (non-renewal and landlord steps), and previously were using type="button" with onClick, but this breaks the ability to use the enter key to proceed and is bad for a11y. So now that we already have the <form> wrapper used on each step with form fields, we can just override the onSubmit instead.

Now just need to make sure the <form> wraps the whole step, and now we can also just give the classname in form for syling and loose the extra div wrapper.