QA fixes to landlord details step (#216)
This fixes the issue of landlord fields getting cleared on "back" from the preview step even if there is no HPD address provided. Now they are only cleared if there is an HPD address to prefill.

I also did a big refactor of all the address verification since we now want to pre-verify and correct hpd provided addresses before the user sees them, and we want to verify/correct addresses even if they are marked "deliverable" without any unit number warnings.

This didn't come up in QA, but if the HPD suggested address has validation errors they were showing immediately on the suggested address page even though the fields aren't visible. This has been fixed by removing shouldValidate option when setting fields. Also fixed issue with missing unit number so that if it isn't present in the hpd address the no-unit checkbox is checked.

Also if you accept the hpd address and click "next" right away if there are validation errors it shows a modal letting you know that the HPD address on record has some missing/invalid fields that need to be manually corrected, and it has a single "edit address" button that will bring up the manual inputs with validation errors, but omitting the "override" infobox warning.

I also updates the styles and changed to title case for the addresses shown for HPD lookup, lob verification confirmation modals, and input fields.

This also adds the Dropdown for State selection, and the conditional validation logic for PR/Urbanization and state 2-char abbreviation.

I've added some sample BBLs for user address for different HPD landlord address options, in local testing you can provide the sampleValues to defaultValues in react-hook-form but for the preview link here are the street addresses to use for the user:

245 Front Street, Brooklyn (bbl: "2022810070", // hpd PR address)
43 Harrison Street, Manhattan (bbl: "1001420025", // hpd landlord complete)
272 Third Avenue, Brooklyn (bbl: "3004400040", // hpd landlord missing zip)
40-38 77th Street, Queens (bbl: "4014870027", // no city or state)
427 80 STREET, Brooklyn (bbl: "3059800077", // no hpd landlord)
We have now removed any notifications to the user that a better address is available from USPS, and instead just automatically accept the corrections from LOB verification. This has the potential to create some confusion for the user since the address they input might not be exactly what appears on the letter and we don't have any possibility for users to force their address exactly as input. But this is also what LOC does, it just doesn't tell them - in the backend after submission when we are sending the address LOB will use their corrected version anyway. Here is an example loom, and the first part where an error is raised is very rare since only the PR/urbanization case could allow for the verification address update to lead to a validation error. But the later example with zipcode and it changing on the letter will probably be common

Note: Now that we call the wow api then the lob verify before loading if there is an HPD landlord address this can take a while to complete, so we will add a loading message. This will be added separately when implementing a similar loading state on the confirmation page.