fix QA issues with landlord address reset and verification (#224)
In QA we noticed that if you start with a user address that has an hpd landlord contact, then go back and change the user address to one that doesn't, when you return to the landlord step the previous landlord values are incorrectly prefilled. This was because in the useEffect that sets the HPD contact as your form fields it only does so if it finds an HPD contact. So if there is no hpd contact found it doesn't reset to default empty values. I've updated it so now it will always reset values on load of the landlord step - which is what we initially intended.

Also, we were getting a default zod error for missing state in landlord address, so I've updated the schema validation to handle undefined.

While fixing the above I noticed a couple other issues.

In rare cases it's possible for the WOW landlord lookup to have allcontacts: null, and this wasn't accounted for so you could be indefinitely stuck on the loading screen for landlord step. This is now fixed with some data validation and early return.

Also I wanted to avoid making the LOB verify api call if we knew that we were missing some required fields. And in the process I realized we were being overly strict in the validation for the LOB verify api. I've relaxed the validation on the tenants2 backend to reflect what the actual LOB API requires (see companion PR: JustFixNYC/tenants2#2495) and updated here accordingly.

I also refactored slightly for some better self-documenting conditional checks.