fix bug preventing deletion of phone input (#226)
There was a very specific bug in the phone number input where under certain conditions it was impossible to fully delete the input.

Details for reproducing error:

fill out the first three digits of the phone number section
click next, which will trigger a form level error
click “back” in the form nav
click “next” in the form nav to go back to contact info page. the 3 digit input should be preserved. now try deleting
Turns out it was related to the parsing of the formatted value returning undefined instead of empty string, so it wasn't getting set in the react-hook-form controlled component field.onChange leaving the last digit there.