Update letter progress bar for semantic html and headers (#246)
We decided to add names for each step as an additional header to the "build your letter" header above the progress bar and we wanted to use the <progress> html element - all to improve screen reader experience.

I've added msg tag step names to the lettersSteps, and added those to the progress bar.

I also was able to adjust just rework the existing progress bar styles to work with the ` element.

Also in the new designs we have an additional progress bar state sfor while the data is loading for the confirmation page and when there is an error. This breaks the general approach of managing the progress bar from the Builder component via letterSteps. So now we leave out the progress bar just for the confirmation and instead add different versions from within the step. This required just a little reworking of styles to move the progress bar within the builder div. Unfortunately having the progress within the step breaks the css transition for just that last step.

Separately we will need to adjust other headers within steps to ensure they all pass a11y check