fix jittery header on scroll (#244)
The top header bar hide as you scroll down and reappears when you scroll up. But it has been really jittery - both in that the page shifts a bit when it appears and that it can disappear/reappear rapidly in reaction to a tiny bit of scrolling, and I believe these two issues combined so that when you're at the bottom of the page the shift and start to affect the scrolling causing wild jittering.

I've reworked how this hide/show behavior is implemented to hopefully be more stable and simple.

The previous useScrollDirection hook is now just useHideHeader returning a boolean. The header uses position fixed instead of sticky and added a margin on body to offset everything to make room, and I've changed to use a 10px threshold on detecting scroll changes, and added a simple sliding in/out transition.

At the bottom of the page things can get jittery, I think because of one or more of these issues:

scrolling past the bottom of the page on mobile
the mobile browser UI appearing/disappearing on scroll as well
I think I might have partially addressed this bottom f page jitter by disabling the show/hide updater when close to the bottom of the page (in practice just requiring a larger sized scroll up to activate the header). Things seem ok to me on safari, but there is still a tiny and inconsistent jitter on ios/chrome.