use loader to restrict/redirect letter step routes (#253)
We want to restrict users from reaching any letter builder step outside of the prescribed progression. So if you are landing directly at /letter/contact_info without having completed the prior steps it should redirect you to /letter. Also once you get to /letter/confirmation you should not be able to go back to any prior steps.

This is now handled with react-router via a Loader and session storage with an array of allowed letter step route names, which we add to from the builder component on "next", and clear when reaching "confirmation".

Also apparently using redirect() like this is preferred by SEO as well over a simple navigate(), since it uses the proper 302 etc.