Minor functional fixes from VQA (#255)
On the frontend we only check that phone numbers are 10 digits, but on the tenants2 backend we also check for area codes starting with 0 or 1, which are invalid. So this updates our frontend validation to do the same, so there's no chance a number like this can get through and cause an error on submit.
On mail choice emails section we automatically disable and uncheck the "CC" checkbox when there is no input for landlord email, but unchecking didn't actually reset the field value to false. So now we use watch and a useEffect to fix this.