Use dynamic heading levels for content box (#260)
In QA we realized we needed to clean up the use of headings to improve screen reader navigation. Some things that should be headings weren't (title of ContentBox and ContentBoxItem) and in some places specific heading levels were hard coded but used on different pages with different heading level contexts breaking the nesting order.

This adds a new dynamic heading component that allows you to set the heading level via prop. Then we use this new Heading for content box title, content box item header, and the internal content box item sections. This allows us to then pass a heading level to the contentbox and then the components can increment the level and pass it down to the children to maintain proper nesting.

I also cleaned up the contentbox component a bit, dropping unused props (step, title vs subttile) from old designs.

I have also made a similar change to the Notice component as well, which was previously using an h3 for the header.

I also updated the styles and fixed some vqa issues with the content of next step items.

For review, this impacts contentboxes and notices on /results, /tenants_rights, /rent_calculator, /letter/next_steps, /letter/allowed_increase, and /letter/confirmation_test.

If this new dynamic heading looks ok we can copy it over to the component library., so I tried to stick to our conventions there with classes and forward ref etc.

Also, to make things a bit more flexible in case you're doing something like we d here with incrementing the header level, if the number goes beyond h6 it changes to a span but keeps the jfcl-heading class