more QA items (#264)
prevent layout shift on reason step select button

add role="alert" to form step global error

I just realized that for some reason voiceover on mac is not announcing the alert here or on the screener survey, but it works on windows NVDA. Apparently VoiceOver and alerts is a common issue, and I think it might be related to focus shifting to the question with the error, which was also a suggested a11y feature. I'll continue looking into this but since this at least fixes windows I'll include this now.

mail choice layout shift issue fix